mjolmacs: remove swift
Since we're writing mjolmacs in Objective-C (for easy integration with
emacs) and passing function pointers around seems cumbersome, let's just
stick with Objective-C for now.

We drop the ShortcutRecorder dependency for a much more simple DDHotKey.
The downside is that it still uses Carbon APIs but if we need to drop
Carbon in a future macOS version, we'll cross that bridge then.