module: refactor NSPanel into using an emacs frame
This is a lot of tedious and hacky code just to get the behaviour of an
NSPanel. First of all, because we're not using the OS native NSPanel, we
need to keep track of the previous application so that we can switch
back to it when our frame is dismissed. This is really hacky and prone
to race conditions / unexpected user behavior.