constants: remove cached emacs_value variables
Upon further inspection, we were getting seemingly random segfaults at
the end of a function call. The first smell was that returning Qnil
sometimes returned "0" instead of nil.

The second smell was Qt causing a segfault. It would seem the randomness
was due to someone (Emacs) doing a garbage collection on these
variables. We were naive to think that we could cache these values.

Lesson learned: never cache anything from emacs_env. It will be marked
for garbage collection once the module calls are finished.